+++ src/GRA.Data/Context.cs
@@ -167,7 +167,6 @@ namespace GRA.Data
         public DbSet<Model.SchoolDistrict> SchoolDistricts { get; set; }
         public DbSet<Model.SchoolType> SchoolTypes { get; set; }
         public DbSet<Model.Site> Sites { get; set; }
-        public DbSet<Model.StaticAvatar> StaticAvatars { get; set; }
         public DbSet<Model.System> Systems { get; set; }
         public DbSet<Model.Trigger> Triggers { get; set; }
         public DbSet<Model.TriggerBadge> TriggerBadges { get; set; }