+++ src/GRA.Domain.Model/Site.cs
@@ -48,7 +48,6 @@ namespace GRA.Domain.Model
         public string FromEmailName { get; set; }
         public string FromEmailAddress { get; set; }
 
-        public bool UseDynamicAvatars { get; set; }
         [MaxLength(100)]
         public string SiteLogoUrl { get; set; }
         [MaxLength(100)]